refactor: extract matches display logic to ViewLeagueMatchesCard component
- Moved all match-related UI and logic from ViewLeagueTeamsCard to a new ViewLeagueMatchesCard component
- Updated ViewLeagueTeamsCard to use the new matches component
- Improved separation of concerns for league teams and matches display
